+++ src/legacy/hooks/usePost.ts
@@ -11,6 +11,8 @@ import {
 } from '../service/post.service'
 import { getUserList } from '../service/user.service'
 
+// IDEA tanstack query로 처리할 시 indalied key로 재요청 처리한다. 그럼 posts를 state로 관리하지 않아도 된다.
+// posts는 서버의 상태이기에 서버 상태관리 라이브러리인 tanstack query를 사용해서 처리한다.
 export const usePost = (limit: number, skip: number) => {
   const [posts, setPosts] = useState<Post[]>([])
   const [loading, setLoading] = useState(false)
