+++ src/pages/PostsManagerPage.tsx
@@ -31,6 +31,7 @@ import { getPostTags } from '../legacy/service/post.service'
 import { deleteComment, getComments, patchComment, postComment, putComment } from '../legacy/service/comments.service'
 import { getUser } from '../legacy/service/user.service'
 import { usePost } from '../legacy/hooks/usePost'
+import { highlightText } from '../legacy/utils/highligtText'
 
 const PostsManager = () => {
   const navigate = useNavigate()
@@ -202,21 +203,6 @@ const PostsManager = () => {
     setSortOrder(params.get('sortOrder') || 'asc')
     setSelectedTag(params.get('tag') || '')
   }, [location.search])
-
-  // 하이라이트 함수 추가
-  const highlightText = (text: string, highlight: string) => {
-    if (!text) return null
-    if (!highlight.trim()) {
-      return <span>{text}</span>
-    }
-    const regex = new RegExp(`(${highlight})`, 'gi')
-    const parts = text.split(regex)
-    return (
-      <span>
-        {parts.map((part, i) => (regex.test(part) ? <mark key={i}>{part}</mark> : <span key={i}>{part}</span>))}
-      </span>
-    )
-  }
 
   // 게시물 테이블 렌더링
   const renderPostTable = () => (
